LibGfx/TIFF: Extract metadata-related definition to their own file
